init evaluator
