--- log.ini
@@ -1,28 +1,0 @@
-[loggers]
-keys=root,torch
-
-[handlers]
-keys=richHandler,nullHandler
-
-[formatters]
-keys=richFormat
-
-[logger_root]
-level=INFO
-handlers=richHandler,nullHandler
-
-[logger_torch]
-level=WARNING
-qualname=torch
-handlers=richHandler
-
-[handler_richHandler]
-class=rich.logging.RichHandler
-
-[handler_nullHandler]  # dummy handler to store formatter object
-class=NullHandler
-formatter=richFormat
-
-[formatter_richFormat]
-format=%(asctime)s %(levelname)-8s %(message)s
-datefmt=%d %b %Y %H:%M:%S
