+++ scripts/example_motion.py
@@ -19,8 +19,10 @@ from trainer.evaluate import Evaluator
 from trainer.models import LSTM
 from trainer.train import Trainer
 
-log_cfg = Path(__file__).parent.resolve().parent / "log.ini"
-logging.config.fileConfig(log_cfg, disable_existing_loggers=False)
+PROJECT_ROOT = Path(__file__).resolve().parent.parent
+logging.config.dictConfig(
+    OmegaConf.to_object(OmegaConf.load(PROJECT_ROOT / "log.yml"))
+)
 logger = logging.getLogger(__name__)
 
 
@@ -58,7 +60,30 @@ def ddp_setup(rank: int, cfg: DictConfig) -> None:
     logger.setLevel(logging.INFO if rank == 0 else logging.WARNING)
 
 
-def main(rank: int, cfg: DictConfig) -> None:
+def initialize_lstm(
+    cfg: DictConfig,
+    dl: torch.utils.data.DataLoader,
+) -> torch.nn.Module:
+    """
+    Create LSTM with hyperparameters from config and input/output size from DL.
+    """
+    return LSTM(
+        input_size=dl.dataset[0][0].shape[-1],
+        out_features=dl.dataset[0][1].data.shape[-1],
+        **cfg.arch.lstm,  # unpack the hyperparams as kwargs
+    )
+
+
+def evaluate(cfg: DictConfig) -> None:
+
+    logger.info("Evaluating")
+    t_dl = get_dl(cfg.test.data.x_dir, cfg.test.data.y_dir)
+    model = initialize_lstm(cfg, t_dl)
+    evaluator = Evaluator(cfg, model, t_dl)
+    evaluator.evaluate(model="checkpoint_best.pth")
+
+
+def train(rank: int, cfg: DictConfig) -> None:
 
     # configure current worker within the DistributedDataParallel context
     if not torch.multiprocessing.current_process().name == "MainProcess":
@@ -73,11 +98,7 @@ def main(rank: int, cfg: DictConfig) -> None:
     )
 
     # instantiate model
-    model = LSTM(
-        input_size=t_dl.dataset[0][0].shape[-1],
-        out_features=t_dl.dataset[0][1].data.shape[-1],
-        **cfg.arch.lstm,  # unpack the rest of the hyperparams as kwargs
-    ).to(rank)
+    model = initialize_lstm(cfg, t_dl).to(rank)
 
     # clone model across all GPUs, if applicable
     if torch.distributed.is_initialized():
@@ -99,17 +120,10 @@ def main(rank: int, cfg: DictConfig) -> None:
     trainer = Trainer(cfg, model, optimizer, t_dl, v_dl, metrics, rank)
     trainer.train()
 
-    # run eval from main process only
-    if rank == 0:
-        logger.info("Evaluating")
-        t_dl = get_dl(cfg.test.data.x_dir, cfg.test.data.y_dir)
-        evaluator = Evaluator(cfg, trainer, t_dl)
-        evaluator.evaluate(model="checkpoint_best.pth")
-        if torch.distributed.is_initialized():
-            logging.info("All processes completed; releasing resources...")
-
     # terminate DDP worker
     if torch.distributed.is_initialized():
+        if rank == 0:
+            logging.info("All processes completed; releasing resources...")
         destroy_process_group()
 
 
@@ -120,9 +134,12 @@ if __name__ == "__main__":
     try:
 
         if cfg.debug:  # cannot use breakpoints in distributed environment
-            main(rank=cfg.cuda.visible_devices[0], cfg=cfg)
-        else:
-            mp.spawn(main, args=([cfg]), nprocs=cfg.cuda.num_gpus)
+            train(rank=cfg.cuda.visible_devices[0], cfg=cfg)
+        else:  # create a subprocess for each GPU
+            mp.spawn(train, args=([cfg]), nprocs=cfg.cuda.num_gpus)
+
+        # run evaluation from main process only
+        evaluate(cfg=cfg)
 
     except (KeyboardInterrupt, Exception):
         logger.exception("")
